+++ CHANGELOG.md
@@ -7,7 +7,51 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/
 [Semantic Versioning]: https://semver.org/spec/v2.0.0.html
 
-## [4.0.0-alpha.0] - 2025-10-30
+## Unreleased
+
+### Added
+
+- Ability to print config for a file. [#158]
+- Support for `package.json` configuration file. [#245]
+
+[#158]: https://github.com/AdguardTeam/AGLint/issues/158
+[#245]: https://github.com/AdguardTeam/AGLint/issues/245
+
+## [4.0.0-alpha.5] - 2025-11-03
+
+### Added
+
+- More exports.
+
+## [4.0.0-alpha.4] - 2025-11-03
+
+### Added
+
+- Support for NodeNext ESM resolution in `.d.ts` files.
+
+## [4.0.0-alpha.3] - 2025-11-03
+
+### Changed
+
+- Renamed CLI entry point file for clarity.
+
+### Fixed
+
+- Build config.
+
+## [4.0.0-alpha.2] - 2025-11-03
+
+### Changed
+
+- Split exports into `@adguard/aglint/linter` and `@adguard/aglint/cli`.
+
+## [4.0.0-alpha.1] - 2025-10-31
+
+### Fixed
+
+- Preset files are now included in the tarball.
+
+## [4.0.0-alpha.0] - 2025-10-31
 
 ### Added